Add failing regression test for issue #274.
