+++ cabal2nix.cabal
@@ -104,6 +104,7 @@ library
     , transformers
     , yaml
   exposed-modules:
+      Cabal2nix
       Distribution.Nixpkgs.Fetch
       Distribution.Nixpkgs.Haskell
       Distribution.Nixpkgs.Haskell.BuildInfo
@@ -221,5 +222,8 @@ test-suite spec
     , text
     , transformers
     , yaml
+    , cabal2nix
     , doctest
+    , tasty
+    , tasty-hunit
   default-language: Haskell2010
